requirements.txt
streamlit
pandas
plotly
altair
scipy
sqlalchemy
psycopg2-binary
python-dotenv
pmdarima
statsmodels
folium
streamlit-folium
pyarrow
matplotlib
openpyxl
tqdm
pg8000